Split TreeView and move base class to params package
