+++ silx/gui/plot3d/SFViewParamTree.py
@@ -42,20 +42,15 @@ from silx.gui.plot.Colormap import Colormap
 from silx.gui.widgets.FloatEdit import FloatEdit
 
 from .ScalarFieldView import Isosurface
-from .params import SubjectItem, ColorItem, ColorEditor, DirectionalLightItem
+from . import params
 
 _logger = logging.getLogger(__name__)
-
-
-class ModelColumns(object):
-    NameColumn, ValueColumn, ColumnMax = range(3)
-    ColumnNames = ['Name', 'Value']
 
 
 # View settings ###############################################################
 
 
-class BackgroundColorItem(ColorItem):
+class BackgroundColorItem(params.ColorItem):
     itemName = 'Background'
 
     def setColor(self, color):
@@ -65,7 +60,7 @@ class BackgroundColorItem(ColorItem):
         return self.getSubject().getBackgroundColor()
 
 
-class ForegroundColorItem(ColorItem):
+class ForegroundColorItem(params.ColorItem):
     itemName = 'Foreground'
 
     def setColor(self, color):
@@ -75,7 +70,7 @@ class ForegroundColorItem(ColorItem):
         return self.getSubject().getForegroundColor()
 
 
-class HighlightColorItem(ColorItem):
+class HighlightColorItem(params.ColorItem):
     itemName = 'Highlight'
 
     def setColor(self, color):
@@ -85,7 +80,7 @@ class HighlightColorItem(ColorItem):
         return self.getSubject().getHighlightColor()
 
 
-class BoundingBoxItem(SubjectItem):
+class BoundingBoxItem(params.SubjectItem):
     """Bounding box, axes labels and grid visibility item.
 
     Item is checkable.
@@ -103,7 +98,7 @@ class BoundingBoxItem(SubjectItem):
             self.getSubject().setBoundingBoxVisible(checked)
 
 
-class OrientationIndicatorItem(SubjectItem):
+class OrientationIndicatorItem(params.SubjectItem):
     """Orientation indicator visibility item.
 
     Item is checkable.
@@ -142,7 +137,7 @@ class ViewSettingsItem(qt.QStandardItem):
             titleItem.setEditable(False)
             self.appendRow([titleItem, cls(subject)])
 
-        nameItem = DirectionalLightItem(
+        nameItem = params.DirectionalLightItem(
             subject.getPlot3DWidget(), 'Light Direction')
         valueItem = qt.QStandardItem()
         self.appendRow([nameItem, valueItem])
@@ -150,7 +145,7 @@ class ViewSettingsItem(qt.QStandardItem):
 
 # Data information ############################################################
 
-class DataChangedItem(SubjectItem):
+class DataChangedItem(params.SubjectItem):
     """
     Base class for items listening to ScalarFieldView.sigDataChanged
     """
@@ -217,7 +212,7 @@ class DataSetItem(qt.QStandardItem):
 
 # Isosurface ##################################################################
 
-class IsoSurfaceRootItem(SubjectItem):
+class IsoSurfaceRootItem(params.SubjectItem):
     """
     Root (i.e : column index 0) Isosurface item.
     """
@@ -285,7 +280,7 @@ class IsoSurfaceRootItem(SubjectItem):
             self.getSubject().setVisible(checked)
 
 
-class IsoSurfaceLevelItem(SubjectItem):
+class IsoSurfaceLevelItem(params.SubjectItem):
     """
     Base class for the isosurface level items.
     """
@@ -378,7 +373,7 @@ class IsoSurfaceLevelSlider(IsoSurfaceLevelItem):
         return True
 
 
-class IsoSurfaceColorItem(SubjectItem):
+class IsoSurfaceColorItem(params.SubjectItem):
     """
     Isosurface color item.
     """
@@ -389,7 +384,7 @@ class IsoSurfaceColorItem(SubjectItem):
         return self.getSubject().sigColorChanged
 
     def getEditor(self, parent, option, index):
-        editor = ColorEditor(parent)
+        editor = params.ColorEditor(parent)
         color = self.getSubject().getColor()
         color.setAlpha(255)
         editor.setColor(color)
@@ -407,7 +402,7 @@ class IsoSurfaceColorItem(SubjectItem):
         return self.getSubject().getColor()
 
 
-class IsoSurfaceAlphaItem(SubjectItem):
+class IsoSurfaceAlphaItem(params.SubjectItem):
     """
     Isosurface alpha item.
     """
@@ -447,7 +442,7 @@ class IsoSurfaceAlphaItem(SubjectItem):
         return True
 
 
-class IsoSurfaceAlphaLegendItem(SubjectItem):
+class IsoSurfaceAlphaLegendItem(params.SubjectItem):
     """Legend to place under opacity slider"""
 
     editable = False
@@ -466,7 +461,7 @@ class IsoSurfaceAlphaLegendItem(SubjectItem):
         return editor
 
 
-class IsoSurfaceCount(SubjectItem):
+class IsoSurfaceCount(params.SubjectItem):
     """
     Item displaying the number of isosurfaces.
     """
@@ -520,7 +515,7 @@ class IsoSurfaceAddRemoveWidget(qt.QWidget):
         self.sigViewTask.emit('remove_iso')
 
 
-class IsoSurfaceAddRemoveItem(SubjectItem):
+class IsoSurfaceAddRemoveItem(params.SubjectItem):
     """
     Item displaying a simple QToolButton allowing to add an isosurface.
     """
@@ -530,7 +525,7 @@ class IsoSurfaceAddRemoveItem(SubjectItem):
         return IsoSurfaceAddRemoveWidget(parent, self)
 
 
-class IsoSurfaceGroup(SubjectItem):
+class IsoSurfaceGroup(params.SubjectItem):
     """
     Root item for the list of isosurface items.
     """
@@ -582,7 +577,7 @@ class IsoSurfaceGroup(SubjectItem):
 
 # Cutting Plane ###############################################################
 
-class ColormapBase(SubjectItem):
+class ColormapBase(params.SubjectItem):
     """
     Mixin class for colormap items.
     """
@@ -667,7 +662,7 @@ class PlaneMaxRangeItem(ColormapBase):
         return True
 
 
-class PlaneOrientationItem(SubjectItem):
+class PlaneOrientationItem(params.SubjectItem):
     """
     Plane orientation item.
     Editor is a QComboBox.
@@ -723,7 +718,7 @@ class PlaneOrientationItem(SubjectItem):
         return True
 
 
-class PlaneInterpolationItem(SubjectItem):
+class PlaneInterpolationItem(params.SubjectItem):
     """Toggle cut plane interpolation method: nearest or linear.
 
     Item is checkable
@@ -752,7 +747,7 @@ class PlaneInterpolationItem(SubjectItem):
         self.getSubject().getCutPlanes()[0].setInterpolation(interpolation)
 
 
-class PlaneDisplayBelowMinItem(SubjectItem):
+class PlaneDisplayBelowMinItem(params.SubjectItem):
     """Toggle whether to display or not values <= colormap min of the cut plane
 
     Item is checkable
@@ -908,7 +903,7 @@ class NormalizationNode(ColormapBase):
         return self.getSubject().getCutPlanes()[0].getColormap().getNormalization()
 
 
-class PlaneGroup(SubjectItem):
+class PlaneGroup(params.SubjectItem):
     """
     Root Item for the plane items.
     """
@@ -959,7 +954,7 @@ class PlaneGroup(SubjectItem):
         self.appendRow([nameItem, valueItem])
 
 
-class PlaneVisibleItem(SubjectItem):
+class PlaneVisibleItem(params.SubjectItem):
     """
     Plane visibility item.
     Item is checkable.
@@ -981,76 +976,15 @@ class PlaneVisibleItem(SubjectItem):
 
 # Tree ########################################################################
 
-class ItemDelegate(qt.QStyledItemDelegate):
-    """
-    Delegate for the QTreeView filled with SubjectItems.
-    """
-
-    sigDelegateEvent = qt.Signal(str)
-
-    def __init__(self, parent=None):
-        super(ItemDelegate, self).__init__(parent)
-
-    def createEditor(self, parent, option, index):
-        item = index.model().itemFromIndex(index)
-        if item:
-            if isinstance(item, SubjectItem):
-                editor = item.getEditor(parent, option, index)
-                if editor:
-                    editor.setAutoFillBackground(True)
-                    if hasattr(editor, 'sigViewTask'):
-                        editor.sigViewTask.connect(self.__viewTask)
-                    return editor
-
-        editor = super(ItemDelegate, self).createEditor(parent,
-                                                        option,
-                                                        index)
-        return editor
-
-    def updateEditorGeometry(self, editor, option, index):
-        editor.setGeometry(option.rect)
-
-    def setEditorData(self, editor, index):
-        item = index.model().itemFromIndex(index)
-        if item:
-            if isinstance(item, SubjectItem) and item.setEditorData(editor):
-                return
-        super(ItemDelegate, self).setEditorData(editor, index)
-
-    def setModelData(self, editor, model, index):
-        item = index.model().itemFromIndex(index)
-        if isinstance(item, SubjectItem) and item._setModelData(editor):
-            return
-        super(ItemDelegate, self).setModelData(editor, model, index)
-
-    def __viewTask(self, task):
-        self.sigDelegateEvent.emit(task)
-
-
-class TreeView(qt.QTreeView):
+class TreeView(params.TreeView):
     """
     TreeView displaying the SubjectItems for the ScalarFieldView.
     """
 
     def __init__(self, parent=None):
         super(TreeView, self).__init__(parent)
-        self.__openedIndex = None
-
-        self.setIconSize(qt.QSize(16, 16))
-
-        header = self.header()
-        if hasattr(header, 'setSectionResizeMode'):  # Qt5
-            header.setSectionResizeMode(qt.QHeaderView.ResizeToContents)
-        else:  # Qt4
-            header.setResizeMode(qt.QHeaderView.ResizeToContents)
-
-        delegate = ItemDelegate()
-        self.setItemDelegate(delegate)
+        delegate = self.itemDelegate()
         delegate.sigDelegateEvent.connect(self.__delegateEvent)
-        self.setSelectionBehavior(qt.QAbstractItemView.SelectRows)
-        self.setSelectionMode(qt.QAbstractItemView.SingleSelection)
-
-        self.clicked.connect(self.__clicked)
 
     def setSfView(self, sfView):
         """
@@ -1058,9 +992,7 @@ class TreeView(qt.QTreeView):
 
         :param sfView: A `ScalarFieldView`
         """
-        model = qt.QStandardItemModel()
-        model.setColumnCount(ModelColumns.ColumnMax)
-        model.setHorizontalHeaderLabels(['Name', 'Value'])
+        model = params.TreeViewModelBase()
 
         item = qt.QStandardItem()
         item.setEditable(False)
@@ -1080,122 +1012,6 @@ class TreeView(qt.QTreeView):
 
         self.setModel(model)
 
-    def setModel(self, model):
-        """
-        Reimplementation of the QTreeView.setModel method. It connects the
-        rowsRemoved signal and opens the persistent editors.
-
-        :param qt.QStandardItemModel model: the model
-        """
-
-        prevModel = self.model()
-        if prevModel:
-            self.__openPersistentEditors(qt.QModelIndex(), False)
-            try:
-                prevModel.rowsRemoved.disconnect(self.rowsRemoved)
-            except TypeError:
-                pass
-
-        super(TreeView, self).setModel(model)
-        model.rowsRemoved.connect(self.rowsRemoved)
-        self.__openPersistentEditors(qt.QModelIndex())
-
-    def __openPersistentEditors(self, parent=None, openEditor=True):
-        """
-        Opens or closes the items persistent editors.
-
-        :param qt.QModelIndex parent: starting index, or None if the whole tree
-            is to be considered.
-        :param bool openEditor: True to open the editors, False to close them.
-        """
-        model = self.model()
-
-        if not model:
-            return
-
-        if not parent or not parent.isValid():
-            parent = self.model().invisibleRootItem().index()
-
-        if openEditor:
-            meth = self.openPersistentEditor
-        else:
-            meth = self.closePersistentEditor
-
-        curParent = parent
-        children = [model.index(row, 0, curParent)
-                    for row in range(model.rowCount(curParent))]
-
-        columnCount = model.columnCount()
-
-        while len(children) > 0:
-            curParent = children.pop(-1)
-
-            children.extend([model.index(row, 0, curParent)
-                             for row in range(model.rowCount(curParent))])
-
-            for colIdx in range(columnCount):
-                sibling = model.sibling(curParent.row(),
-                                        colIdx,
-                                        curParent)
-                item = model.itemFromIndex(sibling)
-                if isinstance(item, SubjectItem) and item.persistent:
-                    meth(sibling)
-
-    def rowsAboutToBeRemoved(self, parent, start, end):
-        """
-        Reimplementation of the QTreeView.rowsAboutToBeRemoved. Closes all
-        persistent editors under parent.
-
-        :param qt.QModelIndex parent: Parent index
-        :param int start: Start index from parent index (inclusive)
-        :param int end: End index from parent index (inclusive)
-        """
-        self.__openPersistentEditors(parent, False)
-        super(TreeView, self).rowsAboutToBeRemoved(parent, start, end)
-
-    def rowsRemoved(self, parent, start, end):
-        """
-        Called when QTreeView.rowsRemoved is emitted. Opens all persistent
-        editors under parent.
-
-        :param qt.QModelIndex parent: Parent index
-        :param int start: Start index from parent index (inclusive)
-        :param int end: End index from parent index (inclusive)
-        """
-        super(TreeView, self).rowsRemoved(parent, start, end)
-        self.__openPersistentEditors(parent, True)
-
-    def rowsInserted(self, parent, start, end):
-        """
-        Reimplementation of the QTreeView.rowsInserted. Opens all persistent
-        editors under parent.
-
-        :param qt.QModelIndex parent: Parent index
-        :param int start: Start index from parent index
-        :param int end: End index from parent index
-        """
-        self.__openPersistentEditors(parent, False)
-        super(TreeView, self).rowsInserted(parent, start, end)
-        self.__openPersistentEditors(parent)
-
-    def keyReleaseEvent(self, event):
-        """
-        Reimplementation of the QTreeView.keyReleaseEvent.
-        At the moment only Key_Delete is handled. It calls the selected item's
-        queryRemove method, and deleted the item if needed.
-
-        :param qt.QKeyEvent event: A key event
-        """
-
-        # TODO : better filtering
-        key = event.key()
-        modifiers = event.modifiers()
-
-        if key == qt.Qt.Key_Delete and modifiers == qt.Qt.NoModifier:
-            self.__removeIsosurfaces()
-
-        super(TreeView, self).keyReleaseEvent(event)
-
     def __removeIsosurfaces(self):
         model = self.model()
         selected = self.selectedIndexes()
@@ -1207,7 +1023,7 @@ class TreeView(qt.QTreeView):
         for index in selected:
             leftIndex = model.sibling(index.row(), 0, index)
             leftItem = model.itemFromIndex(leftIndex)
-            if isinstance(leftItem, SubjectItem) and leftItem not in items:
+            if isinstance(leftItem, params.SubjectItem) and leftItem not in items:
                 items.append(leftItem)
 
         isos = [item for item in items if isinstance(item, IsoSurfaceRootItem)]
@@ -1221,17 +1037,6 @@ class TreeView(qt.QTreeView):
                 self,
                 'Remove isosurface',
                 'Select an iso-surface to remove it')
-
-    def __clicked(self, index):
-        """
-        Called when the QTreeView.clicked signal is emitted. Calls the item's
-        leftClick method.
-
-        :param qt.QIndex index: An index
-        """
-        item = self.model().itemFromIndex(index)
-        if isinstance(item, SubjectItem):
-            item.leftClicked()
 
     def __delegateEvent(self, task):
         if task == 'remove_iso':
